+++ tests/elapsed_timer.py
@@ -1,0 +1,37 @@
+"""
+Elapsed time timer.
+"""
+
+from __future__ import absolute_import
+
+import datetime
+
+
+class ElapsedTimer(object):
+    """
+    Elapsed time timer.
+
+    Calculates elapsed time between initiation/reset and access.
+    """
+
+    def __init__(self):
+        """ Initiate the object with current time"""
+        self.start_time = datetime.datetime.now()
+
+    def reset(self):
+        """ Reset the start time for the timer"""
+        self.start_time = datetime.datetime.now()
+
+    def elapsed_ms(self):
+        """ Get the elapsed time in milliseconds. returns floating
+            point representation of elapsed time in seconds.
+        """
+        dt = datetime.datetime.now() - self.start_time
+        return ((dt.days * 24 * 3600) + dt.seconds) * 1000 + \
+            dt.microseconds / 1000.0
+
+    def elapsed_sec(self):
+        """ get the elapsed time in seconds. Returns floating
+            point representation of time in seconds
+        """
+        return self.elapsed_ms() / 1000